backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-google-genai
langgraph
faiss-cpu
numpy
python-dotenv
scikit-learn
sentence-transformers
pypdf